requirements.txt
cycler==0.10.0
matplotlib==2.1.0
numpy==1.13.3
pyenchant==1.6.11
PyMySQL==0.7.11
six==1.11.0
